+++ app/src/main/java/woowacourse/shopping/data/remote/NetworkClient.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.remote.cart
+package woowacourse.shopping.data.remote
 
 import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
 import kotlinx.serialization.json.Json
@@ -6,8 +6,11 @@ import okhttp3.MediaType.Companion.toMediaType
 import okhttp3.OkHttpClient
 import retrofit2.Retrofit
 import woowacourse.shopping.BuildConfig
+import woowacourse.shopping.data.remote.cart.CartService
+import woowacourse.shopping.data.remote.cart.HeaderInterceptor
+import woowacourse.shopping.data.remote.product.ProductService
 
-object CartClient {
+object NetworkClient {
     private val contentType = "application/json".toMediaType()
 
     private val okHttpClient =
@@ -16,15 +19,16 @@ object CartClient {
             .addInterceptor(HeaderInterceptor(BuildConfig.USER_ID, BuildConfig.USER_PASSWORD))
             .build()
 
-    private val _retrofitService: CartService by lazy {
+    private val retrofitService: Retrofit by lazy {
         Retrofit
             .Builder()
             .baseUrl(BuildConfig.BASE_URL)
             .client(okHttpClient)
-            .addConverterFactory(Json.asConverterFactory(contentType))
+            .addConverterFactory(Json.Default.asConverterFactory(contentType))
             .build()
-            .create(CartService::class.java)
     }
 
-    fun getRetrofitService(): CartService = _retrofitService
+    fun getCartService(): CartService = retrofitService.create(CartService::class.java)
+
+    fun getProductService(): ProductService = retrofitService.create(ProductService::class.java)
 }
