+++ app/src/main/java/woowacourse/shopping/data/remote/cart/CartRepository.kt
@@ -1,10 +1,11 @@
 package woowacourse.shopping.data.remote.cart
 
+import woowacourse.shopping.data.remote.NetworkClient
 import woowacourse.shopping.data.remote.cart.CartResponse.Content
 import kotlin.text.substringAfterLast
 
 class CartRepository {
-    suspend fun fetchAllCart() = CartClient.getRetrofitService().requestCart(size = Int.MAX_VALUE)
+    suspend fun fetchAllCart() = NetworkClient.getCartService().requestCart(size = Int.MAX_VALUE)
 
     suspend fun findCartByProductId(productId: Long): Content? {
         val response = fetchAllCart()
@@ -12,11 +13,11 @@ class CartRepository {
         return matched
     }
 
-    suspend fun fetchCart(page: Int) = CartClient.getRetrofitService().requestCart(page = page)
+    suspend fun fetchCart(page: Int) = NetworkClient.getCartService().requestCart(page = page)
 
     suspend fun addToCart(cartRequest: CartRequest): Result<Long> =
         try {
-            val response = CartClient.getRetrofitService().addToCart(cartRequest = cartRequest)
+            val response = NetworkClient.getCartService().addToCart(cartRequest = cartRequest)
             if (response.isSuccessful) {
                 val id =
                     response.headers()["Location"]?.substringAfterLast("/")?.toLongOrNull() ?: 0
@@ -30,7 +31,7 @@ class CartRepository {
 
     suspend fun deleteCart(id: Long): Result<Unit> =
         try {
-            val response = CartClient.getRetrofitService().deleteFromCart(id = id)
+            val response = NetworkClient.getCartService().deleteFromCart(id = id)
             if (response.isSuccessful) {
                 Result.success(Unit)
             } else {
@@ -45,7 +46,7 @@ class CartRepository {
         cartQuantity: CartQuantity,
     ) = try {
         val response =
-            CartClient.getRetrofitService().updateCart(id = id, cartQuantity = cartQuantity)
+            NetworkClient.getCartService().updateCart(id = id, cartQuantity = cartQuantity)
         if (response.isSuccessful) {
             Result.success(Result.success(Unit))
         } else {
@@ -57,7 +58,7 @@ class CartRepository {
 
     suspend fun getCartCounts(): Result<Long> =
         try {
-            val response = CartClient.getRetrofitService().getCartCounts()
+            val response = NetworkClient.getCartService().getCartCounts()
             if (response.isSuccessful) {
                 val totalCount = response.body()?.quantity?.toLong() ?: 0L
                 Result.success(totalCount)
